Rename shared to elements
